Rename Root.tsx to index.tsx
