+++ lab3/src/main.tsx
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom/client';
 import './index.css';
 import { BrowserRouter } from 'react-router-dom';
 import { setTheme } from './themes/themes';
-import Root from './routes/Root';
+import Root from './routes';
 
 setTheme();
